+++ drizzle/0011_confused_wonder_man.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `settings` ADD `suppress_empty_reports` integer DEFAULT true NOT NULL;